+++ Model/SavePlayState.cs
@@ -9,8 +9,7 @@ namespace WinUIMusicPlayer.Model
         public int Id { get; set; } = 1; // 固定 ID 为 1，方便管理
         public PlayMode PlayMode { get; set; }
         public int? LastPlayedMusicId { get; set; }
-        public float Volume { get; set; } = 0.5f;
-
+        public double Volume { get; set; } = 0.5;
         public string sortOrder { get; set; } = "DefaultOrder";
     }
 }
